🚚 Rename: 공지사항, 회원가입, 로그인, 비번 찾기, 프로필 수정을 common으로 이동 (url 변경됨)
