+++ pages/common/auth/changePW.module.scss
@@ -1,4 +1,4 @@
-@import "../../styles/main.module.scss";
+@import "../../../styles/main.module.scss";
 .changePW {
   @extend %page-base;
