+++ pages/common/auth/findPW.module.scss
@@ -1,4 +1,4 @@
-@import "../../styles/main.module.scss";
+@import "../../../styles/main.module.scss";
 .findPWSection {
   @extend %page-base;
   display: flex;
