+++ pages/common/auth/login.module.scss
@@ -1,4 +1,4 @@
-@import "../../styles/main.module.scss";
+@import "../../../styles/main.module.scss";
 .loginSection {
   display: flex;
   flex-direction: column;
