+++ pages/common/auth/signup.module.scss
@@ -1,4 +1,4 @@
-@import "../../styles/main.module.scss";
+@import "../../../styles/main.module.scss";
 .signUp01 {
   @extend %page-base;
   position: relative;
