+++ pages/common/auth/withdraw.module.scss
@@ -1,4 +1,4 @@
-@import "../../styles/main.module.scss";
+@import "../../../styles/main.module.scss";
 .deleteAccount {
   @extend %page-base;
