+++ pages/common/notice/template.module.scss
